requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
pandas>=2.2.2
openpyxl>=3.1.2
playwright>=1.48.0
tqdm>=4.66.4
